Revert "reset solver storage (#1941)"
This reverts commit 76f42becc9177f30057706f091889ace43e84df9.
