+++ .github/workflows/benchmark-build.yaml
@@ -65,11 +65,13 @@ jobs:
             target/release/cli
             target/release/commitment-task
             target/release/submit-transactions
-            target/release/utils
+            target/release/reset-storage
             target/release/deploy
+            target/release/keygen
             target/release/permissionless-builder
             target/release/nasty-client
             target/release/espresso-dev-node
+            target/release/pub-key
             target/release/espresso-bridge
 
   build-arm:
@@ -114,11 +116,13 @@ jobs:
             target/release/cli
             target/release/commitment-task
             target/release/submit-transactions
-            target/release/utils
+            target/release/reset-storage
             target/release/deploy
+            target/release/keygen
             target/release/permissionless-builder
             target/release/nasty-client
             target/release/espresso-dev-node
+            target/release/pub-key
             target/release/espresso-bridge
 
   build-dockers:
