+++ .github/workflows/build.yml
@@ -72,11 +72,13 @@ jobs:
             target/release/cli
             target/release/commitment-task
             target/release/submit-transactions
-            target/release/utils
+            target/release/reset-storage
             target/release/deploy
+            target/release/keygen
             target/release/permissionless-builder
             target/release/nasty-client
             target/release/espresso-dev-node
+            target/release/pub-key
             target/release/espresso-bridge
             target/release/marketplace-solver
             target/release/marketplace-builder
@@ -125,11 +127,13 @@ jobs:
             target/release/cli
             target/release/commitment-task
             target/release/submit-transactions
-            target/release/utils
+            target/release/reset-storage
             target/release/deploy
+            target/release/keygen
             target/release/permissionless-builder
             target/release/nasty-client
             target/release/espresso-dev-node
+            target/release/pub-key
             target/release/espresso-bridge
             target/release/marketplace-solver
             target/release/marketplace-builder
