+++ marketplace-solver/src/state.rs
@@ -326,7 +326,6 @@ impl GlobalState {
             acquire_timeout: None,
             require_ssl: false,
             migrations: true,
-            reset: false,
         };
 
         let client = PostgresClient::connect(opts)
